==> code/analysis/02b_GRCCA-genes/06c_SENS-RCCA/01_characterize-RCCA-res-GSEA.qmd <==
---
title: "02b-06c-01. Assess functional, cell-type, and risk gene enrichments of RCCA results"
author: "Rachel Smith"
format: html
editor: visual
---

Characterize RCCA results using GSEA.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
base_dir <- "~/Documents/PhD/projects/sgacc_wgcna_grcca/"
```

## Setup

```{r load_data}
source(paste0(base_dir, "scripts/load_WGCNA_res.R"))

## Load RCCA data
load(paste0(base_dir, "objects/RCCA_results.Rdata")) # df_lvs, df_y_res, df_x_res
```

## Run GSEA

```{r run_gsea}
### Get ranked gene list ###
my_rcca_ranked <- df_x_res_rcca %>% 
    arrange(-pearsons_r) %>% 
    dplyr::select(ensembl_gene_id, pearsons_r) %>% 
    deframe


### Cell-type GSEA ###
df_rcca_fgsea_cell <- fgsea(pathways = cell_types, 
                          stats = my_rcca_ranked, 
                          eps = 0
)  %>% 
    as_tibble() %>% 
    dplyr::rename("cell_type" = "pathway") %>% 
    arrange(-abs(NES))


### GO GSEA ###
df_rcca_fgsea_go <- fgsea(pathways = go_pathways, 
                        stats = my_rcca_ranked, 
                        eps = 0
) %>% 
    as_tibble() %>% 
    dplyr::rename("go_id" = "pathway") %>% 
    arrange(-abs(NES)) %>% 
    left_join(df_go_terms) %>% # add names of GO terms
    dplyr::select(go_id, term, ontology, everything())


### Benchmark lists ###
df_rcca_fgsea_benchmark <- fgsea(pathways = benchmarking_lists, 
                               stats = my_rcca_ranked, 
                               eps = 0
) %>% 
    as_tibble() %>% 
    dplyr::rename("benchmark_list" = "pathway") %>% 
    arrange(-abs(NES))
```

## Save objects for figures

```{r save}
save(df_rcca_fgsea_cell, df_rcca_fgsea_go, df_rcca_fgsea_benchmark,
     file = paste0(objects_dir, "RCCA_enrichment_benchmark_res.Rdata")
)
```

## Export RCCA GSEA results as supplemental table

```{r tableS10}
write_xlsx(
    list(
        "A.Cell-type" = df_rcca_fgsea_cell %>% 
            mutate(genes = map_chr(leadingEdge, ~ f_collapse_genes(.x, index = 30000))) %>% 
            dplyr::select(-leadingEdge) %>% unnest(cols = c(genes)),
        "B.GO" = df_rcca_fgsea_go %>% 
            dplyr::filter(padj < 0.05) %>%
            mutate(genes = map_chr(leadingEdge, ~ f_collapse_genes(.x, index = 30000))) %>% 
            dplyr::select(-leadingEdge) %>% unnest(cols = c(genes)),
        "C.Risk genes" = df_rcca_fgsea_benchmark %>% 
            mutate(genes = map_chr(leadingEdge, ~ f_collapse_genes(.x, index = 30000))) %>% 
            dplyr::select(-leadingEdge) %>% unnest(cols = c(genes))
    ), 
    path = paste0(tables_dir, "TableS10.RCCA_GSEA_enrichments.xlsx")
)
```

## Plots

### Figure S13D \| RCCA results risk gene enrichment

```{r FigS13d.prep}
df_fig13d <- df_rcca_fgsea_benchmark %>% 
    mutate(class = ifelse(str_detect(benchmark_list, "SCZ"), "SCZ", "other") %>% 
               factor(levels = c("SCZ", "other"))) %>% 
    mutate(benchmark_list = str_remove_all(benchmark_list, "SCZ|\\)|\\(") %>% 
               str_trim %>% 
               str_replace(", ", ",\n")
    ) %>% 
    mutate(label_color = ifelse(padj < 0.05, "white", "black")) %>% 
    mutate(box_color = ifelse(padj < 0.05, "black", "transparent")) %>% 
    mutate(label = round(NES, 2) %>% format(digits = 3))
```

```{r FigS13d.plot}
df_fig13d %>% 
    ggplot(aes(x = benchmark_list, y = 1)) +
    geom_tile(aes(fill = NES, color = I(box_color)), width = 0.95, height = 0.95, linewidth = 0.75) +
    geom_text(aes(label = label, color = I(label_color)), size = 3) +
    facet_wrap(vars(class), scales = "free", nrow = 2) +
    scale_fill_gradient(low = "white", high = "#67001F", limits = c(0.838, 1.58)) +
    guides(fill = guide_colorbar(title.hjust = 0.5, title.position = "left")) +
    labs(x = NULL, y = NULL,
         title = "D | Risk gene enrichment") +
    theme(axis.text.y = element_blank(),
          axis.ticks = element_blank(),
          axis.line = element_blank(),
          axis.text.x = element_text(margin = margin(t = -3)),
          strip.background = element_rect(fill = "white", color = "black"),
          legend.margin = margin(r = -15),
          legend.position = "left",
          legend.justification = "top",
          legend.title = element_text(angle = 90),
          legend.key.height = unit(0.75, "cm"),
          legend.key.width = unit(0.15, "cm")
    )

## Save
map(
  .x = c(".png", ".pdf"),
  .f = ~ ggsave(paste0(supp_figures_dir, "S13D.Risk_gene_enrichment", .x), 
                width = 2.5, height = 2)
)
```
