==> code/analysis/02b_GRCCA-genes/06c_SENS-RCCA/02_compare-GRCCA-RCCA.qmd <==
---
title: "02b-06c-02. Compare RCCA vs GRCCA results"
author: "Rachel Smith"
format: html
editor: visual
---

Compare RCCA vs GRCCA results in the context of the SCZ literature.

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
base_dir <- "~/Documents/PhD/projects/sgacc_wgcna_grcca/"
```

## Setup

```{r load_data}
source(paste0(base_dir, "scripts/load_WGCNA_res.R"))

## Load RCCA results
load(paste0(base_dir, "objects/RCCA_results.Rdata"))
load(paste0(objects_dir, "GRCCA_results.Rdata")) # df_lvs, df_y_res, df_x_res

# Load enrichment results if needed
load(paste0(analysis_objects_dir, "RCCA_enrichment_benchmark_res.Rdata"))
```

## Combine X & Y results in GRCCA & RCCA

```{r combine_results}
## Y
df_y_res_all <- df_y_res_rcca %>% 
    mutate(analysis = "rcca", .before = 1) %>% 
    bind_rows(
        df_y_res %>% 
            mutate(analysis = "grcca", .before = 1)
        
    )

## X
df_x_res_all <- df_x_res_rcca %>% 
    mutate(analysis = "rcca", .before = 1) %>% 
    bind_rows(
        df_x_res %>% 
            mutate(analysis = "grcca", .before = 1)
        
    )
```

## Test hypergeometric overlap of GRCCA and RCCA hits

```{r hypergeometric_test}
### Loop through and run hypergeometric tests across structure correlation thresholds ###

thresholds <- seq(0.01, 0.25, by = 0.01)
df_rcca_grcca_hypergeometric_benchmark <- expand_grid(
    analysis = c("rcca", "grcca"),
    benchmark_list = names(benchmarking_lists),
    threshold = thresholds
) %>% 
    mutate(
        hypergeometric_res = pmap(
            .l = list(..1 = analysis, ..2 = threshold, ..3 = benchmark_list),
            .f = ~ f_hypergeometric(gene_list1 = df_x_res_all %>% 
                                        dplyr::filter(analysis == ..1) %>% 
                                        top_n(n = ..2*length(ensembl_gene_ids), wt = abs(pearsons_r)) %>% 
                                        pull(ensembl_gene_id), 
                                    gene_list2 = benchmarking_lists[[..3]],
                                    gene_universe = ensembl_gene_ids
            )
        ),
        p_value = map(
            .x = hypergeometric_res,
            .f = ~ .x$p_value
        )
    ) %>% 
    unnest(cols = c(p_value)) %>%
    mutate(p_adj = p.adjust(p_value, method = "fdr"))
```

## Save objects for figures

```{r save}
save(df_x_res_all, df_y_res_all, df_rcca_grcca_hypergeometric_benchmark,
     file = paste0(objects_dir, "compare_GRCCA_RCCA.Rdata")
)
```

## Plots

### Figure S11A \| Correlate GRCCA and RCCA covariate (y) structure correlations

```{r FigS11a.plot, fig.width=6.5, fig.height=5}
df_y_res_all %>% 
    pivot_wider(id_cols = covariate, names_from = analysis, values_from = pearsons_r) %>% 
    ggplot(aes(x = rcca, y = grcca)) +
    geom_point(aes(fill = rcca + grcca), shape = 21, size = 2.5) +
    geom_vline(aes(xintercept = 0), color = "gray") +
    geom_hline(aes(yintercept = 0), color = "gray") +
    geom_abline(lty = 2) +
    geom_text_repel(aes(label = covariate), size = 2.5, max.overlaps = 20, box.padding = 1.25,
                     nudge_y = 0.3, force = 150) +
    scale_fill_gradientn(colors = gene_weight_color_scale, limits = c(-1.84, 1.84), guide = "none") +
    labs(title = "A | Covariate structure correlations",
         x = "RCCA", y = "GRCCA")

map(
  .x = c(".png", ".pdf"),
  .f = ~ ggsave(paste0(supp_figures_dir, "S11A.covariate_structCor_correlation", .x), 
                width = 6.5, height = 5)
)
```

### Figure S11B \| Correlate GRCCA and RCCA gene (x) structure correlations

```{r FigS11b.plot, fig.width=6.5, fig.height=5}
df_x_res_all %>% 
    pivot_wider(id_cols = ensembl_gene_id, names_from = analysis, values_from = pearsons_r) %>% 
    ggplot(aes(x = rcca, y = grcca)) +
    geom_point(aes(color = rcca + grcca), size = 1) +
    geom_vline(aes(xintercept = 0), color = "gray") +
    geom_hline(aes(yintercept = 0), color = "gray") +
    geom_abline(lty = 2) +
    stat_cor(label.sep = "\n") +
    #geom_label_repel(aes(label = gene_symbol), size = 3, max.overlaps = 50, box.padding = 1.5) +
    scale_color_gradientn(colors = gene_weight_color_scale, 
                          limits = c(-1.14, 1.14), guide = "none") +
    labs(title = "B | Gene structure correlations",
         x = "RCCA", y = "GRCCA")

map(
  .x = c(".png", ".pdf"),
  .f = ~ ggsave(paste0(supp_figures_dir, "S11B.gene_structCor_correlation", .x), 
                width = 6.5, height = 5)
)
```

### Figure S11C \| Venn diagram of significant gene overlap

```{r FigS11c.prep}
## Venn diagram (overlap with Nirmala's results)
sig_grcca_genes <- df_x_res %>%
    dplyr::filter(p_adj < 0.05 & abs(z_score) > 2) %>%
    pull(ensembl_gene_id)
sig_rcca_genes <- df_x_res_rcca %>% 
    dplyr::filter(p_adj < 0.05 & abs(z_score) > 2) %>% 
    pull(ensembl_gene_id)

length(sig_grcca_genes) 
length(sig_rcca_genes) 

hypergeometric_overlap <- f_hypergeometric(sig_grcca_genes, sig_rcca_genes, gene_universe = ensembl_gene_ids)
length(hypergeometric_overlap$overlap_genes) 
# Note: Venn diagram is typically made in vector graphics software or using ggvenn/VennDiagram package.
# Printing stats for reference.
print(paste("Overlap:", length(hypergeometric_overlap$overlap_genes)))
print(paste("P-value:", hypergeometric_overlap$p_value))
```

### Figure S11D \| RCCA results risk gene enrichment

```{r FigS11d.prep}
df_fig13d <- df_rcca_fgsea_benchmark %>% 
    mutate(class = ifelse(str_detect(benchmark_list, "SCZ"), "SCZ", "other") %>% 
               factor(levels = c("SCZ", "other"))) %>% 
    mutate(benchmark_list = str_remove_all(benchmark_list, "SCZ|\\)|\\(") %>% 
               str_trim %>% 
               str_replace(", ", ",\n")
    ) %>% 
    mutate(label_color = ifelse(padj < 0.05, "white", "black")) %>% 
    mutate(box_color = ifelse(padj < 0.05, "black", "transparent")) %>% 
    mutate(label = round(NES, 2) %>% format(digits = 3))
```

```{r FigS11d.plot, fig.width=6.5, fig.height=5}
df_fig13d %>% 
    ggplot(aes(x = benchmark_list, y = 1)) +
    geom_tile(aes(fill = NES, color = I(box_color)), 
              width = 0.95, height = 0.95, linewidth = 0.75) +
    geom_text(aes(label = label, color = I(label_color)), size = 3) +
    facet_wrap(vars(class), scales = "free", nrow = 2) +
    scale_fill_gradient(low = "white", high = "#67001F", limits = c(0.838, 1.58)) +
    guides(fill = guide_colorbar(title.hjust = 0.5, title.position = "left")) +
    labs(x = NULL, y = NULL,
         title = "D | Risk gene enrichment") +
    theme(axis.text.y = element_blank(),
          axis.ticks = element_blank(),
          axis.line = element_blank(),
          axis.text.x = element_text(margin = margin(t = -3)),
          strip.background = element_rect(fill = "white", color = "black"),
          legend.margin = margin(r = -15),
          legend.position = "left",
          legend.justification = "top",
          legend.title = element_text(angle = 90),
          legend.key.height = unit(0.75, "cm"),
          legend.key.width = unit(0.15, "cm")
    )

map(
  .x = c(".png", ".pdf"),
  .f = ~ ggsave(paste0(supp_figures_dir, "S11D.Risk_gene_enrichment", .x), 
                width = 6.5, height = 5)
)
```
